Se actualizan componentes

+++ src/test/java/com/devops/dxc/devops/DevopsApplicationTests.java
@@ -24,17 +24,17 @@ class DevopsApplicationTests {
 	@Test
 	void testRetiroMaximo(){
 		Dxc diezxciento = ufService.obtenerDiezxCiento(70000000, 0);
-		int uf = ufService.validaUf().getValor();
-		assertEquals("Retiro máximo", (uf * 150), diezxciento.getDxc());
+		double uf = ufService.validaUf().getValor();
+		assertEquals("Retiro máximo", (int)(uf * 150), diezxciento.getDxc());
 	}
 
 	@Test
 	void testSaldoRetiroMaximo(){
 		Dxc diezxciento = ufService.obtenerDiezxCiento(70000000, 0);
 		diezxciento.getDxc();
-		int uf = ufService.validaUf().getValor();
+		double uf = ufService.validaUf().getValor();
 
-		assertEquals("Saldo retiro máximo", (70000000-(uf * 150)), diezxciento.getSaldo());
+		assertEquals("Saldo retiro máximo", (70000000-(int)(uf * 150)), diezxciento.getSaldo());
 	}
 
 	@Test
@@ -81,7 +81,7 @@ class DevopsApplicationTests {
 
 	@Test
 	void testUFValida() {
-		int uf = ufService.validaUf().getValor();
+		double uf = ufService.validaUf().getValor();
 		assertTrue("UF mayor a 29000", uf > 29000);
 	}
 
@@ -93,43 +93,43 @@ class DevopsApplicationTests {
 
 	@Test
 	void testImpuesto2() {	// 669,910 - 1,488,690
-		int uf = ufService.validaUf().getValor();
+		double uf = ufService.validaUf().getValor();
 		Dxc diezxciento = ufService.obtenerDiezxCiento(50000000, 1200000);
-		assertEquals("Impuesto 0.04 segundo rango de sueldos",(int)(0.04*150*uf), diezxciento.getImpuesto());
+		assertEquals("Impuesto 0.04 segundo rango de sueldos",(int)(0.04 * (int)(150*uf)), diezxciento.getImpuesto());
 	}
 
 	@Test
 	void testImpuesto3() {	// 1,488,690 - 2,481,150
-		int uf = ufService.validaUf().getValor();
+		double uf = ufService.validaUf().getValor();
 		Dxc diezxciento = ufService.obtenerDiezxCiento(50000000, 1700000);
-		assertEquals("Impuesto 0.08 tercer rango de sueldos",(int)(0.08*150*uf), diezxciento.getImpuesto());
+		assertEquals("Impuesto 0.08 tercer rango de sueldos",(int)(0.08 * (int)(150*uf)), diezxciento.getImpuesto());
 	}
 
 	@Test
 	void testImpuesto4() {	// 2,481,151 - 3,466,667
-		int uf = ufService.validaUf().getValor();
+		double uf = ufService.validaUf().getValor();
 		Dxc diezxciento = ufService.obtenerDiezxCiento(50000000, 2500000);
-		assertEquals("Impuesto 0.135 cuarto rango de sueldos",(int)(0.135*150*uf), diezxciento.getImpuesto());
+		assertEquals("Impuesto 0.135 cuarto rango de sueldos",(int)(0.135 * (int)(150*uf)), diezxciento.getImpuesto());
 	}
 
 	@Test
 	void testImpuesto5() {	// 3,466,667 - 4,458,334
-		int uf = ufService.validaUf().getValor();
+		double uf = ufService.validaUf().getValor();
 		Dxc diezxciento = ufService.obtenerDiezxCiento(50000000, 4200000);
-		assertEquals("Impuesto 0.23 quinto rango de sueldos",(int)(0.23*150*uf), diezxciento.getImpuesto());
+		assertEquals("Impuesto 0.23 quinto rango de sueldos",(int)(0.23 * (int)(150*uf)), diezxciento.getImpuesto());
 	}
 
 	@Test
 	void testImpuesto6() {	// 4,458,333.4 - 5,950,000
-		int uf = ufService.validaUf().getValor();
+		double uf = ufService.validaUf().getValor();
 		Dxc diezxciento = ufService.obtenerDiezxCiento(50000000, 4700000);
-		assertEquals("Impuesto 0.304 sexto rango de sueldos",(int)(0.304*150*uf), diezxciento.getImpuesto());
+		assertEquals("Impuesto 0.304 sexto rango de sueldos",(int)(0.304 * (int)(150*uf)), diezxciento.getImpuesto());
 	}
 
 	@Test
 	void testImpuesto7() {	// > 5,950,000
-		int uf = ufService.validaUf().getValor();
+		double uf = ufService.validaUf().getValor();
 		Dxc diezxciento = ufService.obtenerDiezxCiento(50000000, 6500000);
-		assertEquals("Impuesto 0.35 septimo rango de sueldos",(int)(0.35*150*uf), diezxciento.getImpuesto());
+		assertEquals("Impuesto 0.35 septimo rango de sueldos",(int)(0.35 * (int)(150*uf)), diezxciento.getImpuesto());
 	}
 }
